+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- **BEHAVIOR** `Camera::mouse_inside()` will now return false for `Bin`'s that have `pass_events` set to `Some(true)`.
+
 # Version 0.14.0 (January 5th, 2022)
 
 - **BREAKING** Update dependency `vulkano` & `vulkano-shaders` to `0.28`.